create product redux
